Removed Turning, Upgraded Player powerups
Turning was a nightmare, would have quickly become spaghetti code, with
tuning needed on each map/track.
instead, we need to use bools, and rotations and math, abstracting it,
as was done for powerups.
Player powerups are easy to understand and use, as shown by the quick
and easy addition of the speed boost