Balance Changes
Player can now Left click to slow the game down. We should either make
this a limited resource, or scale the points. The speed of the game has
been capped at 1.5x, instead of 1x. Maybe the faster a player plays, and
with fewer slo-mo uses, the higher their point value.